Redesign Calendar with month snapshot, ledger grid, and branded day sheet.
Aligns the Calendar tab with the E+F mockup using serene forest-green styling, progress bars per day, and a half-height day sheet.

+++ src/utils/dateUtils.ts
@@ -54,8 +54,8 @@ export function formatMemberSince(value: string): string {
 
 export function formatShortDate(date: string): string {
   return parseLocalDate(date).toLocaleDateString(undefined, {
-    weekday: 'short',
-    month: 'short',
+    weekday: 'long',
+    month: 'long',
     day: 'numeric',
   })
 }
